+++ src/utils/mining.ts
@@ -60,10 +60,10 @@ export const chainReadyForNewBlock = (
   chainStatus: ChainStatus,
   newDataTimestamp: number,
 ): [ready: boolean, error: string | undefined] => {
-  const deltaT = timestamp() - chainStatus.lastDataTimestamp - chainStatus.timePadding;
+  const deltaT = newDataTimestamp - chainStatus.lastDataTimestamp - chainStatus.timePadding;
 
   if (deltaT < 0) {
-    return [false, `skipping ${chainStatus.nextBlockId.toString()}: waiting for next round T${deltaT}`];
+    return [false, `skipping ${newDataTimestamp}: waiting for next round T${deltaT}`];
   }
 
   if (newDataTimestamp <= chainStatus.lastDataTimestamp) {